refactor(editor): extract OverlayInput as editor-ui component
